+++ packages/block-logs-stream/src/groupLogsByBlockNumber.ts
@@ -1,11 +1,11 @@
-import { BlockNumber, Log } from "viem";
-import { NonPendingLog, isNonPendingLog } from "./isNonPendingLog";
+import { BlockNumber } from "viem";
 import { bigIntSort, isDefined } from "@latticexyz/common/utils";
-import { debug } from "./debug";
 
-export type GroupLogsByBlockNumberResult<TLog extends Log> = {
+type PartialLog = { blockNumber: bigint; logIndex: number };
+
+export type GroupLogsByBlockNumberResult<TLog extends PartialLog> = {
   blockNumber: BlockNumber;
-  logs: readonly NonPendingLog<TLog>[];
+  logs: TLog[];
 }[];
 
 /**
@@ -22,25 +22,16 @@ export type GroupLogsByBlockNumberResult<TLog extends Log> = {
  * @returns An array of objects where each object represents a distinct block and includes the block number,
  * the block hash, and an array of logs for that block.
  */
-export function groupLogsByBlockNumber<TLog extends Log>(
+export function groupLogsByBlockNumber<TLog extends PartialLog>(
   logs: readonly TLog[],
   toBlock?: BlockNumber
 ): GroupLogsByBlockNumberResult<TLog> {
-  // Pending logs don't have block numbers, so filter them out.
-  const nonPendingLogs = logs.filter(isNonPendingLog);
-  if (logs.length !== nonPendingLogs.length) {
-    debug(
-      "pending logs discarded",
-      logs.filter((log) => !isNonPendingLog(log))
-    );
-  }
-
-  const blockNumbers = Array.from(new Set(nonPendingLogs.map((log) => log.blockNumber)));
+  const blockNumbers = Array.from(new Set(logs.map((log) => log.blockNumber)));
   blockNumbers.sort(bigIntSort);
 
   const groupedBlocks = blockNumbers
     .map((blockNumber) => {
-      const blockLogs = nonPendingLogs.filter((log) => log.blockNumber === blockNumber);
+      const blockLogs = logs.filter((log) => log.blockNumber === blockNumber);
       if (!blockLogs.length) return;
       blockLogs.sort((a, b) => (a.logIndex < b.logIndex ? -1 : a.logIndex > b.logIndex ? 1 : 0));