+++ packages/cli/src/commands/dev-contracts.ts
@@ -6,7 +6,6 @@ import { loadConfig, resolveConfigPath } from "@latticexyz/config/node";
 import { StoreConfig } from "@latticexyz/store";
 import { tablegen } from "@latticexyz/store/codegen";
 import path from "path";
-import { tsgen } from "../render-ts";
 import { debounce } from "throttle-debounce";
 import { worldgenHandler } from "./worldgen";
 import { WorldConfig } from "@latticexyz/world";
@@ -17,7 +16,6 @@ import { rmSync } from "fs";
 type Options = {
   rpc?: string;
   configPath?: string;
-  tsgenOutput: string;
 };
 
 const commandModule: CommandModule<Options, Options> = {
@@ -35,15 +33,10 @@ const commandModule: CommandModule<Options, Options> = {
         type: "string",
         decs: "Path to MUD config",
       },
-      tsgenOutput: { type: "string", demandOption: true, desc: "Directory to output MUD typescript definition files" },
     });
   },
 
   async handler(args) {
-    if (!args.tsgenOutputDir) {
-      console.error("No output provided");
-    }
-
     // Initial cleanup
     await forge(["clean"]);
 
@@ -137,9 +130,6 @@ const commandModule: CommandModule<Options, Options> = {
       // Run tablegen to generate tables based on the config
       const outPath = path.join(srcDirectory, config.codegenDirectory);
       await tablegen(config, outPath);
-
-      // Run tsgen to regenerate recs types based on the mud config
-      await tsgen(config, args.tsgenOutput);
     }
 
     /** Codegen to run if contracts changed */