+++ packages/common/src/chains/mudFoundry.ts
@@ -1,0 +1,9 @@
+import { foundry } from "viem/chains";
+import { MUDChain } from "./types";
+
+export const mudFoundry = {
+  ...foundry,
+  fees: {
+    defaultPriorityFee: 0n,
+  },
+} as const satisfies MUDChain;