+++ packages/react/src/index.ts
@@ -2,4 +2,5 @@ export * from "./useComponentValue";
 export * from "./useDeprecatedComputedValue";
 export * from "./useEntityQuery";
 export * from "./useObservableValue";
+export * from "./usePromise";
 export * from "./store-cache";