+++ packages/store-sync/src/blockLogsToStorage.test.ts
@@ -2,7 +2,7 @@ import { beforeEach, describe, expect, it, vi } from "vitest";
 import { BlockLogsToStorageOptions, blockLogsToStorage } from "./blockLogsToStorage";
 import storeConfig from "@latticexyz/store/mud.config";
 import { isDefined } from "@latticexyz/common/utils";
-import { TableId } from "@latticexyz/common";
+import { TableId } from "@latticexyz/common/deprecated";
 
 const mockedCallbacks = {
   registerTables: vi.fn<
@@ -120,6 +120,7 @@ describe("blockLogsToStorage", () => {
               "blockNumber": 5448n,
               "operations": [
                 {
+                  "address": "0x5FbDB2315678afecb367f032d93F642f64180aa3",
                   "fieldName": "amount",
                   "fieldValue": 8,
                   "key": {
@@ -173,6 +174,7 @@ describe("blockLogsToStorage", () => {
         "blockNumber": 5448n,
         "operations": [
           {
+            "address": "0x5FbDB2315678afecb367f032d93F642f64180aa3",
             "fieldName": "amount",
             "fieldValue": 8,
             "key": {