+++ packages/store-sync/src/blockLogsToStorage.ts
@@ -7,7 +7,7 @@ import {
   TableSchema,
 } from "@latticexyz/protocol-parser";
 import { StoreConfig, ConfigToKeyPrimitives as Key, ConfigToValuePrimitives as Value } from "@latticexyz/store";
-import { TableId } from "@latticexyz/common";
+import { TableId } from "@latticexyz/common/deprecated";
 import { Address, Hex, decodeAbiParameters, getAddress, parseAbiParameters } from "viem";
 import { debug } from "./debug";
 import { isDefined } from "@latticexyz/common/utils";
@@ -202,6 +202,7 @@ export function blockLogsToStorage<TConfig extends StoreConfig = StoreConfig>({
           //       they'll eventually be turned into "events", but unclear if that should translate to client storage operations
           return {
             log,
+            address: getAddress(log.address),
             type: "SetRecord",
             ...tableId,
             key,
@@ -217,6 +218,7 @@ export function blockLogsToStorage<TConfig extends StoreConfig = StoreConfig>({
           >[typeof fieldName];
           return {
             log,
+            address: getAddress(log.address),
             type: "SetField",
             ...tableId,
             key,
@@ -228,6 +230,7 @@ export function blockLogsToStorage<TConfig extends StoreConfig = StoreConfig>({
         if (log.eventName === "StoreDeleteRecord") {
           return {
             log,
+            address: getAddress(log.address),
             type: "DeleteRecord",
             ...tableId,
             key,