+++ packages/store-sync/src/trpc-indexer/common.ts
@@ -1,9 +1,12 @@
 import { Hex } from "viem";
-import type { TableRecord } from "@latticexyz/store";
-import type { Table } from "../common";
-
-export type TableWithRecords = Table & { records: TableRecord[] };
+import { TableWithRecords } from "../common";
 
 export type StorageAdapter = {
-  findAll: (chainId: number, address: Hex) => Promise<{ blockNumber: bigint | null; tables: TableWithRecords[] }>;
+  findAll: (
+    chainId: number,
+    address?: Hex
+  ) => Promise<{
+    blockNumber: bigint | null;
+    tables: TableWithRecords[];
+  }>;
 };